fix building

+++ src/client/front_render/d3d11/d3d11_render_manager.cpp
@@ -391,7 +391,7 @@ namespace tc
                 *Occluded = true;
             }
         } else {
-            LOGE("Draw failed: {}", Ret);
+            LOGE("Draw failed: {}", (int)Ret);
         }
 
         return Ret;
